requirements.txt
api4jenkins==1.6
python-dotenv==0.19.0
